+++ src/structures/Interaction.js
@@ -2,6 +2,7 @@
 
 const { Base, Permissions } = require('discord.js');
 const SnowflakeUtil = require('../util/SnowflakeUtil');
+const { InteractionType } = require('discord-api-types/v9');
 
 /**
  * Represents an Interaction.
@@ -65,6 +66,10 @@ class Interaction extends Base {
 
   inRawGuild() {
     return Boolean(this.guildId && !this.guild && this.member);
+  }
+
+  isRepliable() {
+    return ![InteractionType.Ping, InteractionType.ApplicationCommandAutocomplete].includes(this.type);
   }
 }